Partway through the Warhol filter
Create a Warhol filter class. I'm approaching this by creating four smaller copies of an original image and applying a different filter to each of those images and replacing those images on the viewer.
I have some of the framework down and I've figured out how to resize and apply the filters to the new small images. I need to work on moving the images around the view so it displays in a quadrant.